修改抄送人格式

+++ src/main/java/cn/hse/service/impl/DelayToApplyForServiceImpl.java
@@ -46,7 +46,7 @@ public class DelayToApplyForServiceImpl implements DelayToApplyForService{
 		String reqCompleteDate = inputJson.getString("reqCompleteDate");
 		String delayToApplyForDec = inputJson.getString("delayToApplyForDec");
 		String delayToApplyForDate = inputJson.getString("delayToApplyForDate");
-		String delayToApplyForNo = String.valueOf((Math.random()*9+1)*100000);
+		String delayToApplyForNo = String.valueOf((int)((Math.random()*9+1)*100000));
 		//封装延期申请单对象
 		DelayToApplyFor delayToApplyForList=new DelayToApplyFor();
 		delayToApplyForList.setDelayToApplyForDate(delayToApplyForDate);
@@ -157,5 +157,4 @@ public class DelayToApplyForServiceImpl implements DelayToApplyForService{
 		}
 		return ResultUtil.result("0", resultMap, new ArrayList<Map<String, Object>>());
 	}
-
 }
